refactor(config): split datalab.config into appinfo/options/persistence/core
The package was two large modules (config.py, config_options.py,
config_persistence.py) mixing several concerns: application identity and
resource paths, the option model, INI persistence, and the initialization
lifecycle. Split along those lines into four single-responsibility modules
with an explicit, cycle-free dependency order:

- appinfo.py: app name/paths, translation/resource setup, and the plugin-path
  helpers that don't touch the shared Conf singleton. No internal dependency.
- options.py (renamed from config_options.py): the DataLabOptions container.
  Depends on appinfo (APP_NAME is now sourced from there instead of being
  duplicated).
- persistence.py (renamed from config_persistence.py): the INI<->options
  mapping, plus DataLabUserConfig/_LegacyConfigReader/
  atomic_save_configuration/migrate_legacy_configuration, moved here from
  config.py now that appinfo.py exists as a leaf module (this move would have
  reintroduced the cycle removed by the OptionStore change otherwise).
- core.py (renamed from config.py): the initialization lifecycle, the shared
  Conf singleton, and the PlotPy default instances tied to it.

The public facade (datalab.config) is unchanged for application code: every
name it exposed before is still exposed, from its new location. A handful of
white-box unit tests that reached into config.py directly (not through the
facade, to exercise internals with no public API) now reach into core.py
instead.

No change to the on-disk INI format or the public API surface.
