+++ datalab/tests/features/macro/macro_app_test.py
@@ -7,7 +7,7 @@ Macro Panel application test
 
 # guitest: show
 
-from datalab.config import config
+from datalab.config import core as config
 from datalab.gui.main import DLMainWindow
 from datalab.utils import qthelpers as qth
