+++ datalab/tests/features/utilities/memstatus_app_test.py
@@ -10,7 +10,7 @@ import numpy as np
 import psutil
 from sigima.objects import Gauss2DParam
 
-from datalab.config import config
+from datalab.config import core as config
 from datalab.env import execenv
 from datalab.tests import datalab_test_app_context
